refactor(ciphertrust): extract VersionedKeyStore to top-level class
Extracted the inner class VersionedKeyStore from CipherTrustMockServer
to improve readability and maintainability. Also converted KeyMetadata
from a class to a record for more concise code.

This addresses robobario's review feedback about the large inner class.
